Rewrite code for generating ray attacks.
I'd misunderstood what the ray generating code needed to do.
Since every possible permutation of attacks needs to be found, it's
significantly faster just to generate a table of ray attacks on the
fly and do lookups into that. This renders the old code for generating
ray attacks useless, so out it goes. RIP 6 hours of my time.
A logical bug in `movegen.rs` was also fixed - the wrong variable was
being checked in a while statement.