Switch the search and UCI threads.
The search now happens on the main thread. A UCI thread now sends the
commands to the main thread for processing.
This allows the search to take references to the main `Engine` struct
without the compiler complaining about lifetimes, since they're on the
same thread now. Later, all the search threads to do the same, from
within a `thread::scope()`.
The extraneous fields of `SearchInfo` were also removed.