Final cleanup before first release.
Well, I say a cleanup. For the most part it is, but there are a couple
of functional changes.
- Update all outdated/badly worded documentation.
- Simplify some code (e.g. `Board::fmt()`)
- Replace `Board::default()` with just `STARTPOS.parse()`, as it only
  takes a couple of microseconds in the worst-case scenario and allows a
  lot of code to be deleted
- Remove any traces of modifying the fields of `Board` directly
- Replace the `perft` command with `go perft`
- Replace the pawn movegen with a more efficient algorithm
- Make the root search immediately add the first move in the generated
  moves to the movelist: this does nothing at depth > 1 but prevents it
  from returning a null move if depth 1 fails to finish

On the pawn movegen: since pawns have so few available moves and all of
their moves are irreversible, the moves of one type (e.g. captures to
the left) can be generated for all pawns simultaneously and then a move
added for each mpve, instead of looping through all the pawns and
generating the moves for each one. This results in a lot more `for`
loops, but no nested ones, so fewer `for` loops are executed overall.
This does change the order the pawn moves are iterated through, but that
will become irrelevant once I add better move ordering.

Although it wasn't the intention, this brings about 10 elo (+-5)
overall. (SPRT results not saved because gaining elo wasn't the main
purpose.) I'm guessing that's due to the pawn movegen speedup.