Replace a TT cutoff hack with proper normalisation.
To avoid incorrect mate scores, I avoided cutting off from a TT hit if
the TT score was a mate score (since it could be at any height in the
tree). This normalises the stored scores so they will always be from the
correct height when looked up.
This has no effect on >99% of positions but massively helps with mate
finding.

Bench: 23109898