Fix an inconsistency in what was counted as a capture.
The move generator counted en passant and promotion as captures whereas
the movepicker did not. I knew about the bug with en passant but never
got round to fixing it; I completely missed promotion.
This probably gains a small amount of elo but I don't care enough to
SPRT it.

Bench: 21447010