Fix the stalls.
Sometimes, the main search thread would finish before the other search
threads and get joined by the scope of the spawning thread. This
spawning thread would then print the best move even though the search
was still technically going on. There's a chance the GUI would then send
the `position` and `go` commands while the search was still continuing,
which would cause the engine to output nothing (which the GUI
interpreted as a stall).
The fix, funnily enough, was to partially revert `904c818`, which was a
'fix' of interfering search results.

Bench: 5550319