Store the whole TT key.
I finally got a hash collision, resulting in an illegal move and a lost
game. This patch stores the full 64-bit key, which doesn't eliminate
collisions, but it makes them far less likely. If I get another
collision, I'll implement a legality check as well.
Minor version number incremented because this is a functional change
(it halves the TT).

Bench: 5552773