Refactor the higher-level code.
`Engine`, `SearchReferences` and `SearchReport` have all been replaced
with `Worker`. This allows a much more intuitive way of starting the
search and obtaining info from it later (create a worker object, make it
start the search and then query it later) instead of creating some
object, passing it to some random function as a parameter, then reading
another random object which contains much less information than the
search produced. It also allows LazySMP and other search techniques to
be added trivially later. Oh yeah, and it also has a slight speedup in
bench (~2.2%) - but frankly I would still commit this change if the
speed difference were in the opposite direction.

Bench: 11578020