Simplify the last two stages of the movepicker.
If a `None` is returned from `MovePicker::next()`, it will never be
called again. Additionally, we don't need to match against the last
variant of an enum because we already know what variant it is. This
allows the last stage to simply and unconditionally return whatever
`MovePicker::find_next_best()` returns in the `Remaining` stage.

Bench: 13606053