Use the New Type idiom for `Depth` and `Evaluation`.
Additionally, `Depth` is now an `i16` instead of a `u8` and `Evaluation`
is an `i32` instead of an `i16`. This allows for calculations with them
without having to worry about overflows.
The maximum and minimum for them remain the same though. This means that
when storing into the transposition table, they are converted easily
into the smaller, old, type (although they're still wrapped in a
struct).

Bench: 14916764